+++ user.h
@@ -10,10 +10,10 @@ class User{
     private:
         string user_name;
         string password;
-        int size;
+        long size;
         bool is_admin;
     public:
-        User(string _user_name, string _password, int _size, bool _is_admin);
+        User(string _user_name, string _password, long _size, bool _is_admin);
         string get_user_name();
         string get_password();
         int get_size();
